requirements.txt
# Dynamic UAV Path Planning scaffold requirements

numpy>=1.23.0
torch>=2.0.0
gymnasium>=0.29.0
pyyaml>=6.0
tqdm>=4.66.0
matplotlib>=3.7.0
pybullet>=3.2.0
gym-pybullet-drones @ git+https://github.com/utiasDSL/gym-pybullet-drones.git
